Move `Array.from` to inside timeout callback
The HTML collection is live and free to change, so calling `Array.from`
within the timeout callback, instead of before allows the collection to
be updated between setting the timeout and calling the callback.

Additionally, this gives the `.forEach` method more certainty about the
ability to call itself on the elements.